# api_users/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Usuarios — Tienda Muebles",
        "description": "API REST responsable de la autenticación y gestión de usuarios. Utiliza Laravel Sanctum para proteger rutas mediante Bearer tokens con abilities basadas en roles.",
        "contact": {
            "email": "admin@tienda-muebles.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api_users.test",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un usuario con rol 'Cliente'. No requiere autenticación.",
                "operationId": "authRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan",
                                        "maxLength": 30
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "García",
                                        "maxLength": 30
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret",
                                        "minLength": 4
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationError"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario y devuelve un Bearer token Sanctum con abilities según el rol. Rol **Admin** obtiene abilities de gestión de usuarios; cualquier otro rol obtiene abilities de perfil.",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login correcto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login correcto."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123xyz..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "abilities": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "profile:view",
                                                "profile:update"
                                            ]
                                        },
                                        "session_id": {
                                            "type": "string",
                                            "example": "sess_64f0a1b2c3"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales incorrectas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cuenta bloqueada temporalmente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuenta bloqueada temporalmente. Inténtalo de nuevo en 5 minutos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca el token Bearer actual y cierra el log de sesión. Requiere autenticación.",
                "operationId": "authLogout",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "session_id": {
                                        "description": "Opcional — cierra también el log de sesión correspondiente",
                                        "type": "string",
                                        "example": "sess_64f0a..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener perfil del usuario autenticado",
                "description": "Devuelve los datos del usuario cuyo token Bearer se incluye en la cabecera, junto a las abilities del token.",
                "operationId": "authProfile",
                "responses": {
                    "200": {
                        "description": "Perfil del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Juan"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "García"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "juan@example.com"
                                        },
                                        "rol_name": {
                                            "type": "string",
                                            "example": "Cliente"
                                        },
                                        "role_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "abilities": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "profile:view",
                                                "profile:update"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/session/{session_id}": {
            "get": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Verificar sesión activa",
                "description": "Valida un session_id y devuelve los datos del usuario. Endpoint público para consumo interno entre microservicios.",
                "operationId": "sessionCheck",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "description": "ID de sesión generado en el login",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "sess_64f0a1b2c3d4e"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesión válida — datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sesión no válida o expirada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión no válida o expirada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "surname": {
                        "type": "string",
                        "example": "García"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "rol_name": {
                        "type": "string",
                        "example": "Cliente"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The name field is required."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer emitido por /api/login. Inclúyelo en la cabecera: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}